adding sqlchipher
