+++ README.md
@@ -1,2 +1,19 @@
 # electvue-express
 Electron bundling with Vue js + Express as a backend
+
+How to run this project:
+**1. Create .env file**
+```cp .env.example .env```
+
+**2. Install Dependencies**
+```npm install```
+
+**3. Rebuild sqlite3 & sqlchipher for electron project**
+Sqlite3 for electron project must be rebuild.
+```npm run rebuild```
+
+**4. Run the project**
+```npm start```
+
+
+
