+++ server/setup/config.js
@@ -7,6 +7,7 @@ const databaseConfig = {
   password: process.env.DATABASE_PASSWORD,
   params: {
     dialect: process.env.DATABASE_ENGINE,
+    dialectModulePath: '@journeyapps/sqlcipher',
     storage: path.join(__dirname, process.env.DATABASE_STORAGE),
     define: {
       underscore: process.env.DATABASE_FUNCTION_DEFINE
